+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 - Fixed wrong `@param` annotation in `Parameters` (#1614)
 - Make `ZRANDMEMBER` prefixable (#1621)
+- Improve connection handshake by pipelining commands (#1622)
 
 ### Added
 - Added retry support (#1616)